refactor(engine): share the sparse tangent helpers between nonlinear solvers
SparseSymbolicCache / cached_symbolic / tangent_free_sparse /
solve_tangent_sparse move from corotational.rs (with a near-duplicate in
arc_length.rs) into solver/sparse_tangent.rs. The shared
tangent_free_sparse takes the already-extracted free block (the
arc_length signature); corotational's call sites inline the extraction
the old private copy did internally. Pure move, no behavior change.
