+++ engine/src/solver/corotational.rs
@@ -5,7 +5,9 @@ use super::dof::DofNumbering;
 use super::assembly;
 use super::constraints::FreeConstraintSystem;
 use super::linear::SPARSE_THRESHOLD;
-use super::time_integration::MAX_DENSE_FALLBACK_DOFS;
+use super::sparse_tangent::{
+    cached_symbolic, solve_tangent_sparse, tangent_free_sparse, SparseSymbolicCache,
+};
 
 /// Resolve the local-Y reference ONCE from the INITIAL element geometry so the
 /// initial and corotated (deformed) local frames share the same reference.
@@ -165,7 +167,9 @@ pub fn solve_corotational_2d(
                     // iterations (pattern is constant); only the numeric
                     // phase runs here, and the resulting factor is reused
                     // for the rest of the increment.
-                    let k_s_csc = tangent_free_sparse(&k_t, n, nf, &cs);
+                    let free_idx: Vec<usize> = (0..nf).collect();
+                    let k_ff = extract_submatrix(&k_t, n, &free_idx, &free_idx);
+                    let k_s_csc = tangent_free_sparse(&k_ff, nf, &cs);
                     let sym = cached_symbolic(&mut sparse_sym_cache, &k_s_csc);
                     match numeric_cholesky(sym, &k_s_csc) {
                         Some(num) => {
@@ -199,7 +203,9 @@ pub fn solve_corotational_2d(
                     }
                 }
             } else if ns >= SPARSE_THRESHOLD {
-                let k_s_csc = tangent_free_sparse(&k_t, n, nf, &cs);
+                let free_idx: Vec<usize> = (0..nf).collect();
+                let k_ff = extract_submatrix(&k_t, n, &free_idx, &free_idx);
+                let k_s_csc = tangent_free_sparse(&k_ff, nf, &cs);
                 solve_tangent_sparse(&k_s_csc, &r_s, &mut sparse_sym_cache)?
             } else {
                 let free_idx: Vec<usize> = (0..nf).collect();
@@ -681,85 +687,8 @@ fn solve_with_cached_factor(factor: &CachedTangentFactor, r_s: &[f64]) -> Vec<f6
     }
 }
 
-/// Sparse symbolic Cholesky factorization cached across Newton iterations.
-/// The tangent's sparsity pattern is constant within a solve call (structure
-/// and constraints don't change), so the symbolic phase runs once and only
-/// the numeric phase repeats per iteration. `col_ptr`/`row_idx` fingerprint
-/// the pattern the symbolic was built from: if a later tangent's pattern
-/// differs (an entry crossing exactly zero), the symbolic is rebuilt.
-struct SparseSymbolicCache {
-    col_ptr: Vec<usize>,
-    row_idx: Vec<usize>,
-    sym: std::rc::Rc<SymbolicCholesky>,
-}
-
-/// Return the symbolic factorization for `k_s`, building (or rebuilding) it
-/// only when the pattern changed since the last call.
-fn cached_symbolic<'a>(
-    cache: &'a mut Option<SparseSymbolicCache>,
-    k_s: &CscMatrix,
-) -> &'a std::rc::Rc<SymbolicCholesky> {
-    let stale = match cache {
-        Some(c) => c.col_ptr != k_s.col_ptr || c.row_idx != k_s.row_idx,
-        None => true,
-    };
-    if stale {
-        *cache = Some(SparseSymbolicCache {
-            col_ptr: k_s.col_ptr.clone(),
-            row_idx: k_s.row_idx.clone(),
-            sym: std::rc::Rc::new(symbolic_cholesky(k_s)),
-        });
-    }
-    &cache.as_ref().unwrap().sym
-}
-
-/// Extract the free block of the dense tangent and apply constraint
-/// reduction in sparse form (used when ns >= SPARSE_THRESHOLD).
-fn tangent_free_sparse(
-    k_t: &[f64],
-    n: usize,
-    nf: usize,
-    cs: &Option<FreeConstraintSystem>,
-) -> CscMatrix {
-    let free_idx: Vec<usize> = (0..nf).collect();
-    let k_ff = extract_submatrix(k_t, n, &free_idx, &free_idx);
-    let k_ff_csc = CscMatrix::from_dense_symmetric(&k_ff, nf);
-    if let Some(ref cs) = cs {
-        cs.reduce_matrix_sparse(&k_ff_csc)
-    } else {
-        k_ff_csc
-    }
-}
-
-/// Solve K_s * x = r_s with sparse Cholesky, reusing the cached symbolic
-/// factorization for the numeric phase. On non-SPD tangents falls back to
-/// dense LU, size-capped by `MAX_DENSE_FALLBACK_DOFS` — mirrors
-/// `time_integration::solve_spd_sparse_or_dense` /
-/// `factor_effective_stiffness`.
-fn solve_tangent_sparse(
-    k_s: &CscMatrix,
-    r_s: &[f64],
-    cache: &mut Option<SparseSymbolicCache>,
-) -> Result<Vec<f64>, String> {
-    let sym = cached_symbolic(cache, k_s);
-    if let Some(num) = numeric_cholesky(sym, k_s) {
-        return Ok(sparse_cholesky_solve(&num, r_s));
-    }
-    let n = k_s.n;
-    if n > MAX_DENSE_FALLBACK_DOFS {
-        return Err(format!(
-            "Tangent stiffness is not SPD (sparse Cholesky) and the dense LU fallback would need a {n}×{n} dense matrix ({} MB), over the {MAX_DENSE_FALLBACK_DOFS}-DOF ceiling",
-            8 * n * n / 1_000_000
-        ));
-    }
-    let mut k_work = k_s.to_dense_symmetric();
-    let mut r_work = r_s.to_vec();
-    lu_solve(&mut k_work, &mut r_work, n).ok_or_else(|| {
-        "Singular tangent stiffness — structure may be a mechanism \
-         or load increment too large"
-            .to_string()
-    })
-}
+// Sparse symbolic factorization cache and tangent solve helpers now live in
+// `super::sparse_tangent` (shared with arc_length and contact).
 
 fn solve_free_dofs(k_ff: &[f64], r_f: &[f64], nf: usize) -> Result<Vec<f64>, String> {
     let mut k_work = k_ff.to_vec();
@@ -1162,7 +1091,9 @@ pub fn solve_corotational_3d(
                     // iterations (pattern is constant); only the numeric
                     // phase runs here, and the resulting factor is reused
                     // for the rest of the increment.
-                    let k_s_csc = tangent_free_sparse(&k_t, n, nf, &cs);
+                    let free_idx: Vec<usize> = (0..nf).collect();
+                    let k_ff = extract_submatrix(&k_t, n, &free_idx, &free_idx);
+                    let k_s_csc = tangent_free_sparse(&k_ff, nf, &cs);
                     let sym = cached_symbolic(&mut sparse_sym_cache, &k_s_csc);
                     match numeric_cholesky(sym, &k_s_csc) {
                         Some(num) => {
@@ -1190,7 +1121,9 @@ pub fn solve_corotational_3d(
                     }
                 }
             } else if ns >= SPARSE_THRESHOLD {
-                let k_s_csc = tangent_free_sparse(&k_t, n, nf, &cs);
+                let free_idx: Vec<usize> = (0..nf).collect();
+                let k_ff = extract_submatrix(&k_t, n, &free_idx, &free_idx);
+                let k_s_csc = tangent_free_sparse(&k_ff, nf, &cs);
                 solve_tangent_sparse(&k_s_csc, &r_s, &mut sparse_sym_cache)?
             } else {
                 let free_idx: Vec<usize> = (0..nf).collect();
